Fix bug related to reading from file storage
The Ui.echo is called every time handleUserInput method is invoked.
However, the readFromStorage method is only invoked and stored into
the program when users input list.

The bug has to be fixed so that Ui.echo will invoke readFromStorage
method the first time Ui.echo is called regardless of the user input.

Let's move the code fragment with readFromStorage under the "list" case
to the beginning of the Ui.echo method.